Add CMake and clang-cl support on MSVS
This adds the development environment on Windows as follows.
* Visual Studio 2022 v17.11.5 (Community) with clang-cl installed.
* CMake 3.30.5
No WSL is required.

Note that MSVC compiler is not working yet. Only Clang is supported.

Compared with the original Makefile build system, it still lacks of
many features because some external tools cannot be executed on
Windows (outside WSL or MingW).
For example, CMake cannot build ico file from PNF, cannot build
translation files and image resource files.

Type: New Feature

+++ src/bflib_math.c
@@ -690,7 +690,7 @@ static long bitScanReverse(long s)
 {
   unsigned long source = (unsigned long)s;
 #if defined(_MSC_VER)
-    DWORD i;
+    unsigned long i;
     uint8_t success = _BitScanReverse(&i, source);
     return success != 0 ? i : -1;
 #elif defined(__GNUC__)
@@ -831,7 +831,7 @@ long LbDiagonalLength(long a, long b)
     return (tmpval >> 13);
 }
 
-float lerp(float a, float b, float f) 
+float lerp(float a, float b, float f)
 {
     return (a * (1.0 - f)) + (b * f);
 }
